- Fixed some uninitialized structures - Fixed compiler warnings
Pre-ci test passed

git-svn-id: http://svn.eurecom.fr/openair4G/trunk@4388 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair1/SIMULATION/TOOLS/abstraction.c
@@ -187,7 +187,7 @@ double compute_sinr(channel_desc_t *desc,
   return(10*log10(avg_sinr/(nb_rb*2)));
 }
 
-u8 pbch_polynomial_degree;
+int pbch_polynomial_degree;
 double a[7];
 
 void load_pbch_desc(FILE *pbch_file_fd) {
